Refactor to create tree structure with OOP approach
To help create a mental model to approach this problem, the code
has been refactored to use a tree structure to generate the
marshalling function body. An interface named `Renderable` is used
to represent any data type that can be rendered as a series of
lines in the marshal function body. Json types are implemented
in `json.go` and protobuf types are implemented in `protojson.go`.
The idea is to start with a struct representing a Renderable message.
For each field in the message, the kind is determined and wrapped
in the appropriate Renderable. Once the tree is constructed, rendering
the top level `Renderable` will kick off a traversal through the tree,
rendering each component as needed in the correct order.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>